Implement deprecation messages for entire modules.
Use this to deprecate base.target and base.build_file_aliases.
These can be referenced in plugins, hence the deprecation cycle.

A followup change will move all remaining internal references
to the new locations.
