+++ src/python/pants/build_graph/BUILD
@@ -8,16 +8,21 @@ python_library(
   dependencies = [
     '3rdparty/python:six',
     '3rdparty/python/twitter/commons:twitter.common.collections',
+    'src/python/pants/backend/core:wrapped_globs',
     'src/python/pants/base:address',
     'src/python/pants/base:address_lookup_error',
     'src/python/pants/base:addressable',
     'src/python/pants/base:build_file',
-    'src/python/pants/base:build_file_aliases',
+    'src/python/pants/base:build_file_target_factory',
     'src/python/pants/base:build_environment',
+    'src/python/pants/base:fingerprint_strategy',
+    'src/python/pants/base:hash_utils',
     'src/python/pants/base:parse_context',
+    'src/python/pants/base:payload',
     'src/python/pants/base:payload_field',
-    'src/python/pants/base:target',
+    'src/python/pants/base:source_root',
     'src/python/pants/base:target_addressable',
+    'src/python/pants/option',
     'src/python/pants/subsystem',
     'src/python/pants/util:memo',
   ]
